refine the attention file in atom-vllm folder remove the multi inheritance and inline the attention methods
